% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtoolbox-package.R
\docType{package}
\name{adtoolbox-package}
\alias{adtoolbox}
\alias{adtoolbox-package}
\title{adtoolbox: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/adtoolbox}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/adtoolbox/issues}
}

}
\keyword{internal}
