% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-logic.R
\name{augment.logreg}
\alias{augment.logreg}
\title{Augment logic regression}
\usage{
\method{augment}{logreg}(x, data = NULL, ...)
}
\arguments{
\item{x}{description}

\item{data}{description}

\item{...}{others}
}
\description{
Augment logic regression
}
