% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-independence.R
\name{calc_independence}
\alias{calc_independence}
\title{Calculate the independence score}
\usage{
calc_independence(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Calculate the independence score
}
\examples{
library(broom)
step_count |>
  fit_logic_reg(response = unexpect,
                predictors = too_many_high_days:min_day_too_low) |>
    augment() |>
    calc_miscla_rate(unexpect, .fitted) |>
    calc_independence()
}
