% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-miscla-rate.R
\name{calc_miscla_rate}
\alias{calc_miscla_rate}
\title{Calculate the misclassification rate (precision, recall, F1)}
\usage{
calc_miscla_rate(data, pred, true, include_f1 = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{pred}{The column name of the predicted values}

\item{true}{The column name of the true values}

\item{include_f1}{Whether to include F1 score}
}
\description{
Calculate the misclassification rate (precision, recall, F1)
}
\examples{
library(broom)
step_count |>
  fit_logic_reg(response = unexpect,
                predictors = too_many_high_days:min_day_too_low) |>
  augment() |>
  calc_miscla_rate(unexpect, .fitted)
}
