% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-logic.R
\name{fit_logic_reg}
\alias{fit_logic_reg}
\title{Fit logic regression}
\usage{
fit_logic_reg(data, response, predictors, ...)
}
\arguments{
\item{data}{A tibble}

\item{response}{the response variable}

\item{predictors}{the predictors}

\item{...}{others}
}
\description{
Fit logic regression
}
\examples{
library(broom)
step_count |>
   fit_logic_reg(response = unexpect,
                 predictors = too_many_high_days:min_day_too_low) |>
   augment()
}
