% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-miscla-rate.R
\docType{data}
\name{calc_metrics}
\alias{calc_metrics}
\alias{harmonic}
\alias{square_root}
\alias{arithmetic}
\alias{geometric}
\alias{contraharmonic}
\title{Calculate combined metrics from precision, recall and independence}
\format{
An object of class \code{quosure} (inherits from \code{formula}) of length 2.

An object of class \code{quosure} (inherits from \code{formula}) of length 2.

An object of class \code{quosure} (inherits from \code{formula}) of length 2.

An object of class \code{quosure} (inherits from \code{formula}) of length 2.

An object of class \code{quosure} (inherits from \code{formula}) of length 2.
}
\usage{
calc_metrics(
  data,
  metrics = c("harmonic", "square_root", "arithmetic", "geometric", "contraharmonic"),
  include_original = TRUE
)

harmonic

square_root

arithmetic

geometric

contraharmonic
}
\arguments{
\item{data}{A data frame}

\item{metrics}{A character vector of metrics to calculate, including harmonic,
square_root, arithmetic, geometric, contraharmonic means}

\item{include_original}{Whether to include the original columns}
}
\description{
Calculate combined metrics from precision, recall and independence
}
\examples{
library(tibble)
set.seed(123)
dt <- tibble(pred = sample(0:1, 10000, replace = TRUE),
             true = sample(0:1, 10000, replace = TRUE))
dt |> calc_miscla_rate(pred, true) |> dplyr::mutate(independence = 0.8) |> calc_metrics()

}
\keyword{datasets}
