% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-logicreg.R
\name{plot_logicreg}
\alias{plot_logicreg}
\title{Plot the logic regression output}
\usage{
plot_logicreg(x, names)
}
\arguments{
\item{x}{A logic regression model output from [fit_logic_reg] or [LogicReg::logreg()]}

\item{names}{A vector of names for the variables}
}
\description{
Plot the logic regression output
}
\examples{
fit <- step_count |>
  fit_logic_reg(response = unexpect,
                predictors = too_many_high_days:min_day_too_low,
                nleaves = 4, seed = 1234)
plot_logicreg(fit)
}
