% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{step_count}
\alias{step_count}
\title{Simulated step count data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 9 columns.
}
\usage{
step_count
}
\description{
Simulated step count data
}
\keyword{datasets}
