% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-miscla-rate.R
\name{calc_miscla_rate}
\alias{calc_miscla_rate}
\title{Calculate the misclassification rate (precision, recall, F1)}
\usage{
calc_miscla_rate(data, pred, true, include_f1 = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{pred}{The column name of the predicted values}

\item{true}{The column name of the true values}

\item{include_f1}{Whether to include F1 score}
}
\description{
Calculate the misclassification rate (precision, recall, F1)
}
\examples{
library(tibble)
set.seed(123)
dt <- tibble(pred = sample(0:1, 10000, replace = TRUE),
             true = sample(0:1, 10000, replace = TRUE))
dt |> calc_miscla_rate(pred, true, include_f1 = TRUE)
}
